[AIRFLOW-6368][Depends on 6367] Move conf tests from test_core to test_configuration (#6925)
